requirements.txt
Django>=3.0,<4.0
django-gravatar2==1.4.4
djangorestframework==3.12.4
django-bootstrap-v5==1.0.11
django-bootstrap-icons==0.8.3
gunicorn==20.1.0
requests==2.26.0
